Placed all search-related code on home page component and will abstract various code pieces into reusable components if needed
